start by pulling out barchart from actions
